license: use LGPLv2.1 or later (was LGPL (2.1|3.0)-only)
There is currently no LGPLv4, so this change has no effect at the
moment.

In case the LGPLv4 arrives and I am not alive to approve/review it,
the lesser of evils is have give blanket approval of all future LGPL
versions (as published by the FSF).  The worse evil is to be stuck
with a license which cannot guarantee the Free-ness of this project
in the future.

This unfortunately means the FSF can theoretically come out with
license terms I do not agree with, but the LGPLv2.1 and LGPLv3
will always remain an option to all users.